fix: added isUserUnAuthorized

+++ app/api/transaction/[transactionId]/route.ts
@@ -1,8 +1,8 @@
 import { deleteTransaction, fetchTransactionsById, updateTransaction } from "@/server/db/transactions";
+import { internalServerError, isUserUnauthorized, unauthorized } from "@/lib/utils/default-response";
 
 import { get } from "lodash";
 import { getAuthSession } from "@/features/auth/utils";
-import { internalServerError } from "@/lib/utils/default-response";
 import { serializeBigIntValues } from "@/lib/utils/bigInt-serializer";
 
 export const GET = async (req: Request, { params }: { params: { transactionId: string } }) => {
@@ -10,9 +10,9 @@ export const GET = async (req: Request, { params }: { params: { transactionId: s
         const session = await getAuthSession();
         const userId = get(session, "user.id");
 
-        // if (isUserUnauthorized(userId)) {
-        //     return unauthorized;
-        // }
+        if (isUserUnauthorized(userId)) {
+            return unauthorized;
+        }
 
         const trxns = await fetchTransactionsById(userId, params.transactionId);
 
@@ -29,11 +29,11 @@ export const DELETE = async (req: Request, { params }: { params: { transactionId
         const session = await getAuthSession();
         const userId = get(session, "user.id");
 
-        // if (isUserUnauthorized(userId)) {
-        //     return unauthorized;
-        // }
+        if (isUserUnauthorized(userId)) {
+            return unauthorized;
+        }
 
-        const trxns = await deleteTransaction(userId, params.transactionId);
+        await deleteTransaction(userId, params.transactionId);
 
         return Response.json({ message: "Transaction deleted successfully" });
     } catch (error) {
@@ -49,9 +49,9 @@ export const PUT = async (req: Request, { params }: { params: { transactionId: s
         const userId = get(session, "user.id");
         const body = await req.json();
 
-        // if (isUserUnauthorized(userId)) {
-        //     return unauthorized;
-        // }
+        if (isUserUnauthorized(userId)) {
+            return unauthorized;
+        }
 
         const trxns = await updateTransaction(userId, params.transactionId, body);
